requirements-test.txt
# Test Dependencies for Menu Translation Database TDD
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0
httpx==0.25.2
pytest-mock==3.12.0
testcontainers[postgresql]==3.7.1
alembic==1.12.1

# Database dependencies
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Additional testing utilities
freezegun==1.2.2
responses==0.24.1